+++ dUsefulStuff.xcodeproj/ChangeLog.markdown
@@ -5,6 +5,16 @@
 * refactored the DCUIRating code and tests. Created strategies for the various calculations and drawing methods. 
 * Fixed EXC_BAD_ACCESS error in DCDialogs:displayMessage: due to un-required release of the message string.
 * Added optional --debug parameter to the main build script which enables building a debug library.
+* Refactored DCCoreData so that all required objects are now created in the constructor. 
+* Added a delete method to DCCoreData.
+* Removed the did quit message from DCCoreData;
+* Remove DCDialogs:displayMessage: from DCCoreData. Now returns error objects instead.
+* Switched to the OCMock framework after building OCMock as a framework.
+* Fixed silly missing ';' in dealloc macro.
+* Added DCUIRatingDelegate protocol to allow notification of changed value from DCUIRating.
+* Fixed bug where external code setting the value dynamically was not refreshing the display.
+* Updated deployment targets and SDKs to IOS4.2.
+* Updated test to not use a depreciated NSFileManager method.
 
 # Changes Version 0.0.5